+++ test/OpenTelemetry.Resources.Host.Tests/HostDetectorTests.cs
@@ -1,6 +1,8 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+using System.Net;
+using System.Net.NetworkInformation;
 #if NET
 using System.Runtime.InteropServices;
 #endif
@@ -9,6 +11,8 @@ namespace OpenTelemetry.Resources.Host.Tests;
 
 public class HostDetectorTests
 {
+    private const string EnableNetworkAddressesEnvVarName = "OTEL_DOTNET_EXPERIMENTAL_HOST_RESOURCE_ENABLE_NETWORK_ADDRESSES";
+
 #if !NETFRAMEWORK
     private const string MacOSMachineIdOutput = @"+-o J293AP  <class IOPlatformExpertDevice, id 0x100000227, registered, matched,$
         {
@@ -48,6 +52,8 @@ public class HostDetectorTests
     [Fact]
     public void TestHostAttributes()
     {
+        using var environment = EnvironmentVariableScope.Create(EnableNetworkAddressesEnvVarName, null);
+
         var resource = ResourceBuilder.CreateEmpty().AddHostDetector().Build();
 
         Assert.NotNull(resource);
@@ -231,4 +237,74 @@ public class HostDetectorTests
         }
     }
 #endif
+
+    [Theory]
+    [InlineData("true")]
+    [InlineData("True")]
+    [InlineData("TRUE")]
+    public void TestHostNetworkAddressesEnabled(string value)
+    {
+        using var environment = EnvironmentVariableScope.Create(EnableNetworkAddressesEnvVarName, value);
+
+        var resource = ResourceBuilder.CreateEmpty().AddHostDetector().Build();
+
+        var resourceAttributes = resource.Attributes.ToDictionary(x => x.Key, x => x.Value);
+
+        Assert.True(resourceAttributes.ContainsKey("host.ip"), "host.ip should be detected when the flag is set and the host has an eligible network interface.");
+        Assert.True(resourceAttributes.ContainsKey("host.mac"), "host.mac should be detected when the flag is set and the host has an eligible network interface.");
+
+        Assert.NotEmpty(Assert.IsType<string[]>(resourceAttributes["host.ip"]));
+        Assert.NotEmpty(Assert.IsType<string[]>(resourceAttributes["host.mac"]));
+    }
+
+    [Theory]
+    [InlineData(null)]
+    [InlineData(" ")]
+    [InlineData("false")]
+    [InlineData("1")]
+    public void TestHostNetworkAddressesNotEnabled(string? value)
+    {
+        using var environment = EnvironmentVariableScope.Create(EnableNetworkAddressesEnvVarName, value);
+
+        var resource = ResourceBuilder.CreateEmpty().AddHostDetector().Build();
+
+        var resourceAttributes = resource.Attributes.ToDictionary(x => x.Key, x => x.Value);
+
+        Assert.False(resourceAttributes.ContainsKey("host.ip"), "host.ip should not be detected when the flag is not set.");
+        Assert.False(resourceAttributes.ContainsKey("host.mac"), "host.mac should not be detected when the flag is not set.");
+    }
+
+    [Theory]
+    [InlineData(OperationalStatus.Up, NetworkInterfaceType.Ethernet, true)]
+    [InlineData(OperationalStatus.Up, NetworkInterfaceType.Wireless80211, true)]
+    [InlineData(OperationalStatus.Up, NetworkInterfaceType.Unknown, true)]
+    [InlineData(OperationalStatus.Up, NetworkInterfaceType.Tunnel, true)]
+    [InlineData(OperationalStatus.Up, NetworkInterfaceType.Loopback, false)]
+    [InlineData(OperationalStatus.Down, NetworkInterfaceType.Ethernet, false)]
+    [InlineData(OperationalStatus.Dormant, NetworkInterfaceType.Ethernet, false)]
+    public void TestShouldIncludeNetworkInterface(OperationalStatus operationalStatus, NetworkInterfaceType networkInterfaceType, bool expected) =>
+        Assert.Equal(expected, HostDetector.ShouldIncludeNetworkInterface(operationalStatus, networkInterfaceType));
+
+    [Theory]
+    [InlineData("192.0.2.1", true)]
+    [InlineData("2001:db8::1", true)]
+    [InlineData("10.0.0.4", true)]
+    [InlineData("169.255.0.1", true)]
+    [InlineData("127.0.0.1", false)]
+    [InlineData("::1", false)]
+    [InlineData("169.254.0.1", false)]
+    [InlineData("fe80::1", false)]
+    public void TestShouldIncludeIpAddress(string address, bool expected) =>
+        Assert.Equal(expected, HostDetector.ShouldIncludeIpAddress(IPAddress.Parse(address)));
+
+    [Theory]
+    [InlineData(new byte[] { 0xAA, 0xBB, 0xCC, 0xDD, 0xEE, 0xFF }, "AA-BB-CC-DD-EE-FF")]
+    [InlineData(new byte[] { 0x00, 0x11, 0x22, 0x33, 0x44, 0xFF }, "00-11-22-33-44-FF")]
+    [InlineData(new byte[] { 0xAA, 0xBB, 0xCC, 0xDD, 0xEE, 0xFF, 0x00, 0x11 }, "AA-BB-CC-DD-EE-FF-00-11")]
+    public void TestFormatPhysicalAddress(byte[] address, string expected) =>
+        Assert.Equal(expected, HostDetector.FormatPhysicalAddress(new PhysicalAddress(address)));
+
+    [Fact]
+    public void TestFormatPhysicalAddressWithEmptyAddress() =>
+        Assert.Null(HostDetector.FormatPhysicalAddress(PhysicalAddress.None));
 }