+++ src/OpenTelemetry.Resources.Host/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Added the `host.ip` and `host.mac` resource attributes, emitted when
+  `OTEL_DOTNET_EXPERIMENTAL_HOST_RESOURCE_ENABLE_NETWORK_ADDRESSES` is `true`.
+  ([#5172](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5172))
+
 ## 1.18.0-beta.1
 
 Released 2026-Aug-21